+++ src/crawler/sixpm.js
@@ -1,0 +1,4 @@
+
+module.exports = {
+    fetch : require('./default').fetch
+}
